+++ config.js
@@ -11,8 +11,8 @@
 
 var _ = require('lodash')
 var path = require('path')
+var utils = require('./utils')
 var fs = require('fs')
-var findUp = require('find-up')
 var ExtractTextPlugin = require('extract-text-webpack-plugin')
 var webpack = require('webpack')
 var webpackDevMiddleware = require('webpack-dev-middleware')
@@ -43,10 +43,49 @@ runtime.ghPagesEnv = {
 runtime.isDev = runtime.lanyonEnv === 'development'
 runtime.isHotLoading = runtime.isDev && ['serve', 'start'].indexOf(process.argv[2]) !== -1
 
-runtime.projectDir = process.env.LANYON_PROJECT || process.cwd()
+runtime.projectDir = process.env.LANYON_PROJECT || process.env.PWD || process.cwd() // <-- symlinked npm will mess up process.cwd() and point to ~/code/lanyon
 
-runtime.gitRoot = path.dirname(findUp.sync('.git', { cwd: runtime.projectDir }))
-runtime.npmRoot = path.dirname(findUp.sync('package.json', { cwd: runtime.projectDir }))
+runtime.npmRoot = utils.upwardDirContaining('package.json', path.dirname(runtime.projectDir))
+runtime.gitRoot = utils.upwardDirContaining('.git', runtime.npmRoot)
+
+var traverse = process.cwd()
+var parts = traverse.split('/')
+var paths = []
+while (parts.length) {
+  var newParts = parts
+  var ppath = newParts.join('/')
+  if (ppath) {
+    paths.push(ppath)
+  }
+  parts.pop()
+}
+console.log(paths)
+process.exit(0)
+// var curPkgName = 'lanyon'
+
+// for (var i = 0; i < 9; i++) {
+//   console.log(i)
+//   var found = pkgUp.sync(traverse)
+//   if (!found) {
+//     break
+//   }
+//   runtime.npmRoot = path.dirname(found)
+//   curPkgName = path.basename(runtime.npmRoot)
+//   if (curPkgName !== 'lanyon') {
+//     break
+//   }
+//   traverse = path.dirname(traverse)
+// }
+// console.log({npmRoot: runtime.npmRoot})
+
+runtime.gitRoot = path.dirname(findUp.sync('.git', { cwd: runtime.npmRoot }))
+// runtime.npmRoot = path.dirname(findUp.sync('package.json', { cwd: runtime.projectDir }))
+// // if (runtime.npmRoot === __dirname) {
+// //   var attemptHigherNpmRoot = findUp.sync('package.json', { cwd: path.dirname(runtime.projectDir) })
+// //   if (attemptHigherNpmRoot) {
+// //     runtime.npmRoot = path.dirname(attemptHigherNpmRoot)
+// //   }
+// // }
 
 runtime.projectPackageFile = path.join(runtime.npmRoot, 'package.json')
 try {