+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 - [ ] Utilize `passthru` more (vs `spawnSync`)
 - [ ] Add `travis` to Gemfile, and run it from our local shims
 
+# v0.0.27
+
+[Diff](https://github.com/kvz/lanyon/compare/v0.0.26...v0.0.27)
+
+- [x] Added `utils.upwardDirContaining` for definitive pathfinding
+
 # v0.0.26
 
 [Diff](https://github.com/kvz/lanyon/compare/v0.0.25...v0.0.26)